pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support_automation"
version = "0.1.0"
description = "Times of India — Support Automation: ingest feedback, classify, cluster, detect spikes, draft replies, notify stakeholders."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "Internal — Times Internet" }
authors = [{ name = "Times Internet — Support Automation Team" }]

dependencies = [
  # Core
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "pyyaml>=6.0",

  # Persistence
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.1",
  "pgvector>=0.2.5",

  # Web API
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",

  # External integrations (lazy-imported by their adapters)
  "google-api-python-client>=2.120",
  "google-auth>=2.28",
  "google-auth-oauthlib>=1.2",
  "atlassian-python-api>=3.41",
  "requests>=2.31",
  "httpx>=0.27",

  # ML / NLP
  "sentence-transformers>=2.6",
  "fasttext-langid>=0.0.1; platform_system != 'Windows'",
  "numpy>=1.26",

  # CLI
  "typer>=0.12",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "vcrpy>=6.0",
  "ruff>=0.4",
  "mypy>=1.9",
]

[project.scripts]
support-admin = "entrypoints.cli.support_admin_cli:main"
ingest-feedback = "entrypoints.cli.ingest_feedback_cli:main"
classify-feedback = "entrypoints.cli.classify_and_cluster_cli:main"
sync-knowledge = "entrypoints.cli.sync_knowledge_base_cli:main"
draft-replies = "entrypoints.cli.draft_replies_cli:main"
detect-spikes = "entrypoints.cli.detect_spikes_cli:main"
send-digest = "entrypoints.cli.send_digest_cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = [
  "domain*",
  "service_layer*",
  "adapters*",
  "entrypoints*",
]
exclude = ["tests*", "web_ui*", "data_fixtures*", "secrets*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra --strict-markers"
markers = [
  "integration: tests that require Postgres",
  "end_to_end: full-pipeline tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
plugins = ["pydantic.mypy"]
